clean up __init__.py
